SP-359 refactor : domain 디렉터리 구조 변경
