--- src/main/java/com/ludo/study/studymatchingplatform/study/domain/Platform.java
@@ -1,8 +1,0 @@
-package com.ludo.study.studymatchingplatform.study.domain;
-
-public enum Platform {
-
-	GATHER,
-	GOOGLE_MEET
-
-}
