--- src/main/java/com/ludo/study/studymatchingplatform/study/domain/Role.java
@@ -1,8 +1,0 @@
-package com.ludo.study.studymatchingplatform.study.domain;
-
-public enum Role {
-
-	OWNER,
-	MEMBER
-
-}
