--- src/main/java/com/ludo/study/studymatchingplatform/study/domain/recruitment/ApplicantStatus.java
@@ -1,11 +1,0 @@
-package com.ludo.study.studymatchingplatform.study.domain.recruitment;
-
-public enum ApplicantStatus {
-
-	UNCHECKED,
-	ACCEPTED,
-	REJECTED,
-	REMOVED,
-	CANCELLED;
-
-}
