+++ src/main/java/com/ludo/study/studymatchingplatform/study/domain/recruitment/applicant/ApplicantStatus.java
@@ -1,0 +1,11 @@
+package com.ludo.study.studymatchingplatform.study.domain.recruitment.applicant;
+
+public enum ApplicantStatus {
+
+	UNCHECKED,
+	ACCEPTED,
+	REFUSED,
+	REMOVED,
+	CANCELLED;
+
+}
