+++ src/main/java/com/ludo/study/studymatchingplatform/study/domain/recruitment/position/Position.java
@@ -1,9 +1,4 @@
-package com.ludo.study.studymatchingplatform.study.domain;
-
-import static jakarta.persistence.FetchType.*;
-
-import java.util.ArrayList;
-import java.util.List;
+package com.ludo.study.studymatchingplatform.study.domain.recruitment.position;
 
 import com.ludo.study.studymatchingplatform.common.entity.BaseEntity;
 
@@ -12,10 +7,8 @@ import jakarta.persistence.Entity;
 import jakarta.persistence.GeneratedValue;
 import jakarta.persistence.GenerationType;
 import jakarta.persistence.Id;
-import jakarta.persistence.OneToMany;
 import lombok.AccessLevel;
 import lombok.AllArgsConstructor;
-import lombok.Builder;
 import lombok.Getter;
 import lombok.NoArgsConstructor;
 import lombok.experimental.SuperBuilder;
@@ -25,24 +18,14 @@ import lombok.experimental.SuperBuilder;
 @SuperBuilder
 @NoArgsConstructor(access = AccessLevel.PROTECTED)
 @AllArgsConstructor
-public class Category extends BaseEntity {
+public class Position extends BaseEntity {
 
 	@Id
 	@GeneratedValue(strategy = GenerationType.IDENTITY)
-	@Column(name = "category_id")
+	@Column(name = "position_id")
 	private Long id;
 
-	@Column(
-			nullable = false,
-			columnDefinition = "char(50)"
-	)
+	@Column(nullable = false, columnDefinition = "char(50)")
 	private String name;
-
-	@OneToMany(
-			mappedBy = "category",
-			fetch = LAZY
-	)
-	@Builder.Default
-	private List<Study> studies = new ArrayList<>();
 
 }
