+++ src/main/java/com/ludo/study/studymatchingplatform/study/domain/study/Platform.java
@@ -1,0 +1,8 @@
+package com.ludo.study.studymatchingplatform.study.domain.study;
+
+public enum Platform {
+
+	GATHER,
+	GOOGLE_MEET
+
+}
