+++ src/main/java/com/ludo/study/studymatchingplatform/study/domain/study/StudyStatus.java
@@ -1,4 +1,4 @@
-package com.ludo.study.studymatchingplatform.study.domain;
+package com.ludo.study.studymatchingplatform.study.domain.study;
 
 public enum StudyStatus {
