+++ src/main/java/com/ludo/study/studymatchingplatform/study/domain/study/participant/Role.java
@@ -1,0 +1,8 @@
+package com.ludo.study.studymatchingplatform.study.domain.study.participant;
+
+public enum Role {
+
+	OWNER,
+	MEMBER
+
+}
